.github/scripts/requirements.txt
PyGithub>=2.0.0
PyYAML>=6.0
